app/src/main/java/com/example/ligaappapi/view/leagues/LeaguesPresenter.kt
package com.example.ligaappapi.view.leagues

import com.example.ligaappapi.api.ApiRepository
import com.example.ligaappapi.api.TheSportDBApi
import com.example.ligaappapi.model.LeagueResponse
import com.example.ligaappapi.util.CoroutineContextProvider
import com.google.gson.Gson
import kotlinx.coroutines.GlobalScope
import kotlinx.coroutines.launch

class LeaguesPresenter (private val view: LeaguesView,
                        private val apiRepository: ApiRepository,
                        private val gson: Gson,
                        private val context: CoroutineContextProvider = CoroutineContextProvider()){


    fun getLeagueList(){
        view.showLoading()
        GlobalScope.launch(context.main) {
            val data = gson.fromJson(apiRepository
                .doRequest(TheSportDBApi.getLeague()).await(),
                LeagueResponse::class.java)

            view.hideLoading()
            view.showLeagueList(data.league)
        }
    }
}